check lag and fix mysql reader offset panic
